@@ -11,7 +11,7 @@ home-page = "https://github.com/orsinium-labs/flake8-todos"
 license = "BSD 3-Clause"
 keywords = "flake8 flake8-plugin linter styleguide code quality"
 requires = ["flake8", "pycodestyle"]
-requires-python = ">=3.6"
+requires-python = ">=3.8"
 description-file = "README.md"
 classifiers = [
     "Development Status :: 4 - Beta", "Environment :: Console",
@@ -28,3 +28,20 @@ dev = ["isort[pyproject]", "pytest"]
 
 [tool.flit.entrypoints."flake8.extension"]
 T00 = "flake8_todos:Checker"
+
+[tool.mypy]
+files = ["mypy_baseline"]
+python_version = 3.8
+ignore_missing_imports = true
+# follow_imports = "silent"
+show_error_codes = true
+check_untyped_defs = true
+no_implicit_optional = true
+strict_equality = true
+warn_redundant_casts = true
+warn_unused_ignores = true
+
+[tool.isort]
+profile = "django"
+lines_after_imports = 2
+skip = ".venvs/"
