@@ -1,14 +1,14 @@
-# built-in
+from __future__ import annotations
+
 from functools import partial
 from tokenize import NAME, generate_tokens
 
-# project
 from flake8_todos import Token
 
 
-def test_token_fields():
+def test_token_fields() -> None:
     getter = partial(next, iter(['a = "b"']))
-    tokens = [Token(token) for token in generate_tokens(getter)]
+    tokens = [Token(token) for token in generate_tokens(getter)]  # type: ignore[arg-type]
     assert len(tokens) == 5
 
     assert tokens[0].type == NAME
